Catch and warn about division by zero at runtime
Consider the following example:

         package main
         func main() {

                number := 1 / 0

                if number == 17 {
                        print("Weird\n")
                }
         }

Without any changes this could be built, and executed, like so:

         $ go run . < examples/divide.go  > build/divide.asm
         $ nasm -felf64 -o build/divide.o build/divide.asm
         $ ld -o build/divide build/divide.o
         $ ./build/divide
         Floating point exception

The floatin-point exception here comes from the attempted divison
by zero.  I did consider catching this at code-generation time, however
that would only work for the case of `foo/0`, rather than:

        a := 3
        b := 0
        c := a / b

So this pull-request adds an explicit handler to print a fatal error
and terminate, and compares the divisor with zero - jumping to that
error-handler if it is found.